# Convert std::unordered_set to v8::Set via native APIs