This API didn’t exist when we originally wrote the native implementation of this module.

The benchmarks with 3 trials show a modest improvement over the previous solution that involved invoking methods on the JS objects from C++. About 7% faster for my tests in splices. Not that great honestly but it's still nice to use the proper API now that it exists.

We should probably perform a native trace if we want more performance, but I suspect populating v8 sets in the actual algorithm rather than bailing everything over in the wrapper might yield better performance. We probably have bigger fish to fry right now however.

Before this PR:

inserts: 186.765ms, 181.718ms, 178.908ms
rangeQueries: 754.014ms, 756.123ms, 741.803ms
splices: 6121.944ms, 6330.046ms, 5959.000ms
deletes: 64.560ms, 68.375ms, 66.130ms

After this PR:

inserts: 174.896ms, 173.341ms, 167.472ms
rangeQueries: 712.952ms, 730.542ms, 711.279ms
splices: 5635.677ms, 5804.875ms, 5739.785ms -- 5726.778
deletes: 70.311ms, 66.241ms, 65.933ms

/cc @as-cii